+++ frontend/src/metabase/static-viz/components/WaterfallChart/stories-data/time-x-scale-two-bars-without-total.json
@@ -1,0 +1,225 @@
+{
+  "card": {
+    "original_card_id": 53,
+    "public_uuid": null,
+    "parameter_usage_count": 0,
+    "created_at": "2024-02-29T02:00:50.35842Z",
+    "dataset": false,
+    "parameters": [],
+    "metabase_version": "v0.48.1-SNAPSHOT (c375960)",
+    "collection": {
+      "metabase.models.collection.root/is-root?": true,
+      "authority_level": null,
+      "name": "Our analytics",
+      "is_personal": false,
+      "id": "root",
+      "can_write": true
+    },
+    "visualization_settings": {
+      "waterfall.show_total": false,
+      "graph.dimensions": ["X"],
+      "graph.metrics": ["Y"]
+    },
+    "collection_preview": true,
+    "entity_id": "DBg238cK-X-ASFlSLei5K",
+    "display": "waterfall",
+    "parameter_mappings": [],
+    "id": 53,
+    "dataset_query": {
+      "database": 1,
+      "type": "native",
+      "native": {
+        "template-tags": {},
+        "query": "select '2022-01-01' x, 10 y\nunion all select '2022-01-02', 10"
+      }
+    },
+    "cache_ttl": null,
+    "embedding_params": null,
+    "made_public_by_id": null,
+    "updated_at": "2024-02-29T03:18:01.885241Z",
+    "moderation_reviews": [],
+    "creator_id": 1,
+    "average_query_time": 40.689655172413794,
+    "type": "question",
+    "dashboard_count": 0,
+    "last_query_start": "2024-02-29T03:18:01.870123Z",
+    "name": "waterfall corrupted",
+    "query_type": "native",
+    "collection_id": null,
+    "enable_embedding": false,
+    "database_id": 1,
+    "can_write": true,
+    "initially_published_at": null,
+    "result_metadata": [
+      {
+        "display_name": "X",
+        "field_ref": [
+          "field",
+          "X",
+          {
+            "base-type": "type/Text"
+          }
+        ],
+        "name": "X",
+        "base_type": "type/Text",
+        "effective_type": "type/Text",
+        "semantic_type": null,
+        "fingerprint": {
+          "global": {
+            "distinct-count": 2,
+            "nil%": 0
+          },
+          "type": {
+            "type/Text": {
+              "percent-json": 0,
+              "percent-url": 0,
+              "percent-email": 0,
+              "percent-state": 0,
+              "average-length": 10
+            }
+          }
+        }
+      },
+      {
+        "display_name": "Y",
+        "field_ref": [
+          "field",
+          "Y",
+          {
+            "base-type": "type/Integer"
+          }
+        ],
+        "name": "Y",
+        "base_type": "type/Integer",
+        "effective_type": "type/Integer",
+        "semantic_type": null,
+        "fingerprint": {
+          "global": {
+            "distinct-count": 1,
+            "nil%": 0
+          },
+          "type": {
+            "type/Number": {
+              "min": 10,
+              "q1": 10,
+              "q3": 10,
+              "max": 10,
+              "sd": 0,
+              "avg": 10
+            }
+          }
+        }
+      }
+    ],
+    "table_id": null,
+    "collection_position": null,
+    "archived": false,
+    "description": null,
+    "displayIsLocked": true
+  },
+  "data": {
+    "rows": [
+      ["2022-01-01", 10],
+      ["2022-01-02", 10]
+    ],
+    "cols": [
+      {
+        "display_name": "X",
+        "source": "native",
+        "field_ref": [
+          "field",
+          "X",
+          {
+            "base-type": "type/Text"
+          }
+        ],
+        "name": "X",
+        "base_type": "type/Text",
+        "effective_type": "type/Text"
+      },
+      {
+        "display_name": "Y",
+        "source": "native",
+        "field_ref": [
+          "field",
+          "Y",
+          {
+            "base-type": "type/Integer"
+          }
+        ],
+        "name": "Y",
+        "base_type": "type/Integer",
+        "effective_type": "type/Integer"
+      }
+    ],
+    "native_form": {
+      "query": "select '2022-01-01' x, 10 y\nunion all select '2022-01-02', 10",
+      "params": null
+    },
+    "results_timezone": "America/Montevideo",
+    "results_metadata": {
+      "columns": [
+        {
+          "display_name": "X",
+          "field_ref": [
+            "field",
+            "X",
+            {
+              "base-type": "type/Text"
+            }
+          ],
+          "name": "X",
+          "base_type": "type/Text",
+          "effective_type": "type/Text",
+          "semantic_type": null,
+          "fingerprint": {
+            "global": {
+              "distinct-count": 2,
+              "nil%": 0
+            },
+            "type": {
+              "type/Text": {
+                "percent-json": 0,
+                "percent-url": 0,
+                "percent-email": 0,
+                "percent-state": 0,
+                "average-length": 10
+              }
+            }
+          }
+        },
+        {
+          "display_name": "Y",
+          "field_ref": [
+            "field",
+            "Y",
+            {
+              "base-type": "type/Integer"
+            }
+          ],
+          "name": "Y",
+          "base_type": "type/Integer",
+          "effective_type": "type/Integer",
+          "semantic_type": null,
+          "fingerprint": {
+            "global": {
+              "distinct-count": 1,
+              "nil%": 0
+            },
+            "type": {
+              "type/Number": {
+                "min": 10,
+                "q1": 10,
+                "q3": 10,
+                "max": 10,
+                "sd": 0,
+                "avg": 10
+              }
+            }
+          }
+        }
+      ]
+    },
+    "insights": null
+  }
+}
