+++ frontend/src/metabase/visualizations/lib/renderer_utils.js
@@ -4,21 +4,15 @@ import _ from "underscore";
 import { getIn } from "icepick";
 
 import { parseTimestamp } from "metabase/lib/time";
-import {
-  NULL_NUMERIC_VALUE,
-  TOTAL_ORDINAL_VALUE,
-} from "metabase/lib/constants";
+import { NULL_NUMERIC_VALUE } from "metabase/lib/constants";
 import { formatNullable } from "metabase/lib/formatting/nullable";
 import { datasetContainsNoResults } from "metabase-lib/queries/utils/dataset";
 
 import {
-  computeTimeseriesDataInverval,
   dimensionIsTimeseries,
   dimensionIsExplicitTimeseries,
-  getTimezone,
-  minTimeseriesUnit,
 } from "./timeseries";
-import { computeNumericDataInverval, dimensionIsNumeric } from "./numeric";
+import { dimensionIsNumeric } from "./numeric";
 
 import { getAvailableCanvasWidth, getAvailableCanvasHeight } from "./utils";
 import { invalidDateWarning, nullDimensionWarning } from "./warnings";
@@ -241,116 +235,6 @@ function parseTimestampAndWarn(value, unit) {
   }
   m.toString = moment_fast_toString;
   return { parsedValue: m };
-}
-
-/*
-
-A waterfall chart is essentially a stacked bar chart.
-It consists of the following (from the topmost):
-- the "total bar" which has a value only for the last one
-- the "green bar" for the positives/increases
-- the "red bar" for the negatives/decreases
-- the "invisible beam" supporting either the green or red bar
-
-Note the green and red bars are mutually exclusive (i.e. if one has
-a positive value, the other is zero). This is done so we need not have
-conditional fill color.
-
-*/
-
-export function syntheticStackedBarsForWaterfallChart(
-  datas,
-  settings = {},
-  series = [],
-) {
-  const showTotal = settings && settings["waterfall.show_total"];
-  const stackedBarsDatas = datas.slice();
-  const mainSeries = stackedBarsDatas[0];
-  const mainValues = mainSeries.map(d => d[1]);
-  const totalValue = mainValues.reduce((total, value) => total + value, 0);
-
-  const { beams } = mainValues.reduce(
-    (t, value) => {
-      return {
-        beams: [...t.beams, t.offset],
-        offset: t.offset + value,
-      };
-    },
-    { beams: [], offset: 0 },
-  );
-
-  const values = mainValues.slice();
-  if (showTotal) {
-    const total = [xValueForWaterfallTotal({ settings, series }), totalValue];
-    if (mainSeries[0]._origin) {
-      // cloning for the total bar
-      total._origin = {
-        seriesIndex: mainSeries[0]._origin.seriesIndex,
-        rowIndex: mainSeries.length,
-        cols: mainSeries[0]._origin.cols,
-        row: total,
-      };
-    }
-    stackedBarsDatas[0] = [...mainSeries, total];
-    // The last one is the total bar, anchor it at 0
-    beams.push(0);
-    values.push(0);
-  } else {
-    stackedBarsDatas[0] = mainSeries;
-  }
-  stackedBarsDatas.push(stackedBarsDatas[0].map(k => k.slice())); // negatives
-  stackedBarsDatas.push(stackedBarsDatas[0].map(k => k.slice())); // positives
-  stackedBarsDatas.push(stackedBarsDatas[0].map(k => k.slice())); // total
-  for (let k = 0; k < values.length; ++k) {
-    stackedBarsDatas[0][k]._waterfallValue = stackedBarsDatas[0][k][1];
-    stackedBarsDatas[0][k][1] = beams[k];
-    stackedBarsDatas[1][k][1] = values[k] < 0 ? values[k] : 0;
-    stackedBarsDatas[2][k][1] = values[k] > 0 ? values[k] : 0;
-    stackedBarsDatas[3][k][1] =
-      !showTotal || k < values.length - 1 ? 0 : totalValue;
-  }
-
-  return stackedBarsDatas;
-}
-
-export function getXInterval({ settings, series }, xValues, warn) {
-  if (isTimeseries(settings)) {
-    // We need three pieces of information to define a timeseries range:
-    // 1. interval - it's really the "unit": month, day, etc
-    // 2. count - how many intervals per tick?
-    // 3. timezone - what timezone are values in? days vary in length by timezone
-    const unit = minTimeseriesUnit(series.map(s => s.data.cols[0].unit));
-    const timezone = getTimezone(series, warn);
-    const { count, interval } = computeTimeseriesDataInverval(xValues, unit);
-    return { count, interval, timezone };
-  } else if (isQuantitative(settings) || isHistogram(settings)) {
-    // Get the bin width from binning_info, if available
-    // TODO: multiseries?
-    const binningInfo =
-      getFirstNonEmptySeries(series).data.cols[0].binning_info;
-    if (binningInfo) {
-      return binningInfo.bin_width;
-    }
-
-    // Otherwise try to infer from the X values
-    return computeNumericDataInverval(xValues);
-  }
-}
-
-export function xValueForWaterfallTotal({ settings, series }) {
-  const xValues = getXValues({ settings, series });
-  const xInterval = getXInterval({ settings, series }, xValues, () => {});
-
-  if (isTimeseries(settings)) {
-    const { count, interval } = xInterval;
-    const lastXValue = xValues[xValues.length - 1];
-    return lastXValue.clone().add(count, interval);
-  } else if (isQuantitative(settings) || isHistogram(settings)) {
-    const maxXValue = _.max(xValues);
-    return maxXValue + xInterval;
-  }
-
-  return TOTAL_ORDINAL_VALUE;
 }
 
 const uniqueCards = series => _.uniq(series.map(({ card }) => card.id)).length;
